refactor(monitoring/alerts): split into env + rules submodules
Cycle 36 LOC reduction: alerts.rs (314) -> mod.rs (110) + rules.rs (188) + env.rs (24).
